windows round 4: LLP64-proof test fixtures
- BoundsRule kAlloc/kCopy + nothrow-new proto: __SIZE_TYPE__ (clang
  parses every snippet, so it expands to the TARGET's size_t; plain
  'unsigned long' is 32-bit on LLP64 Windows and the extent modeling
  rightly refuses the mismatched size type -> missed-finding repro'd
  and fixed locally under -target x86_64-pc-windows-msvc)
- IntOverflow SixtyFourBitProductReported: long long (64-bit intent
  must survive LLP64)
- BrokenTu/McpServer fixtures: generic_string() paths — raw C:\ paths
  spliced into hand-built JSON are invalid escapes

+++ tests/IntOverflowRuleTest.cpp
@@ -179,13 +179,15 @@ TEST(IntOverflowRuleTest, StaticHelperBoundedButSafeClean) {
 TEST(IntOverflowRuleTest, SixtyFourBitProductReported) {
     // v0.4 recall round: a 64-bit product still collapses to top() in
     // the int64-based interval math, so the site is proved from the
-    // OPERAND intervals via __int128 corner arithmetic instead
+    // OPERAND intervals via checked int64 corner arithmetic instead
     // (evalEscapes64) — this pinned FN became a finding deliberately.
+    // `long long`, not `long`: the 64-bit intent must survive LLP64
+    // Windows, where long is 32 bits.
     IntOverflowRule rule;
     auto results = runRule(rule, R"(
-        long f() {
-            long a = 100000000000L;
-            long b = 100000000000L;
+        long long f() {
+            long long a = 100000000000LL;
+            long long b = 100000000000LL;
             return a * b;
         }
     )");
